+++ Sources/Sentry/SentryCoreDataTracker.m
@@ -39,7 +39,7 @@
 {
     __block id<SentrySpan> fetchSpan;
     [SentrySDK.currentHub.scope useSpan:^(id<SentrySpan> _Nullable span) {
-        fetchSpan = [span startChildWithOperation:SENTRY_COREDATA_FETCH_OPERATION
+        fetchSpan = [span startChildWithOperation:SentrySpanOperation.coredataFetchOperation
                                       description:[self descriptionFromRequest:request]];
         fetchSpan.origin = SentryTraceOriginAutoDBCoreData;
     }];
@@ -77,7 +77,7 @@
             [self groupEntitiesOperations:context];
 
         [SentrySDK.currentHub.scope useSpan:^(id<SentrySpan> _Nullable span) {
-            saveSpan = [span startChildWithOperation:SENTRY_COREDATA_SAVE_OPERATION
+            saveSpan = [span startChildWithOperation:SentrySpanOperation.coredataSaveOperation
                                          description:[self descriptionForOperations:operations
                                                                           inContext:context]];
             saveSpan.origin = SentryTraceOriginAutoDBCoreData;