+++ Sources/Sentry/SentryUIViewControllerPerformanceTracker.m
@@ -14,7 +14,6 @@
 #    import "SentryTraceOrigins.h"
 #    import "SentryTracer.h"
 #    import <SentryInAppLogic.h>
-#    import <SentrySpanOperations.h>
 #    import <UIKit/UIKit.h>
 #    import <objc/runtime.h>
 
@@ -129,7 +128,7 @@
         NSString *name = [SwiftDescriptor getViewControllerClassName:controller];
         spanId = [self.tracker startSpanWithName:name
                                       nameSource:kSentryTransactionNameSourceComponent
-                                       operation:SentrySpanOperationUILoad
+                                       operation:SentrySpanOperation.uiLoad
                                           origin:SentryTraceOriginAutoUIViewController];
 
         // Use the target itself to store the spanId to avoid using a global mapper.
@@ -202,7 +201,7 @@
             SENTRY_LOG_DEBUG(@"Tracking UIViewController.viewWillAppear");
             [self.tracker measureSpanWithDescription:@"viewWillAppear"
                                           nameSource:kSentryTransactionNameSourceComponent
-                                           operation:SentrySpanOperationUILoad
+                                           operation:SentrySpanOperation.uiLoad
                                               origin:SentryTraceOriginAutoUIViewController
                                              inBlock:callbackToOrigin];
         };
@@ -268,7 +267,7 @@
         void (^duringBlock)(void) = ^{
             [self.tracker measureSpanWithDescription:lifecycleMethod
                                           nameSource:kSentryTransactionNameSourceComponent
-                                           operation:SentrySpanOperationUILoad
+                                           operation:SentrySpanOperation.uiLoad
                                               origin:SentryTraceOriginAutoUIViewController
                                              inBlock:callbackToOrigin];
         };
@@ -312,14 +311,14 @@
         void (^duringBlock)(void) = ^{
             [self.tracker measureSpanWithDescription:@"viewWillLayoutSubviews"
                                           nameSource:kSentryTransactionNameSourceComponent
-                                           operation:SentrySpanOperationUILoad
+                                           operation:SentrySpanOperation.uiLoad
                                               origin:SentryTraceOriginAutoUIViewController
                                              inBlock:callbackToOrigin];
 
             SentrySpanId *layoutSubViewId =
                 [self.tracker startSpanWithName:@"layoutSubViews"
                                      nameSource:kSentryTransactionNameSourceComponent
-                                      operation:SentrySpanOperationUILoad
+                                      operation:SentrySpanOperation.uiLoad
                                          origin:SentryTraceOriginAutoUIViewController];
 
             objc_setAssociatedObject(controller,
@@ -368,7 +367,7 @@
 
             [self.tracker measureSpanWithDescription:@"viewDidLayoutSubviews"
                                           nameSource:kSentryTransactionNameSourceComponent
-                                           operation:SentrySpanOperationUILoad
+                                           operation:SentrySpanOperation.uiLoad
                                               origin:SentryTraceOriginAutoUIViewController
                                              inBlock:callbackToOrigin];
 
@@ -430,7 +429,7 @@
     } else {
         [self.tracker measureSpanWithDescription:description
                                       nameSource:kSentryTransactionNameSourceComponent
-                                       operation:SentrySpanOperationUILoad
+                                       operation:SentrySpanOperation.uiLoad
                                           origin:SentryTraceOriginAutoUIViewController
                                     parentSpanId:spanId
                                          inBlock:callbackToOrigin];