+++ Sources/SentrySwiftUI/SentryTracedView.swift
@@ -43,7 +43,7 @@ public struct SentryTracedView<Content: View>: View {
     let content: () -> Content
     let name: String
     let nameSource: SentryTransactionNameSource
-    
+
     let traceOrigin = "auto.ui.swift_ui"
 
     public init(_ viewName: String? = nil, content: @escaping () -> Content) {
@@ -70,7 +70,12 @@ public struct SentryTracedView<Content: View>: View {
         var transactionCreated = false
         if SentryPerformanceTracker.shared.activeSpanId() == nil {
             transactionCreated = true
-            let transactionId = SentryPerformanceTracker.shared.startSpan(withName: self.name, nameSource: nameSource, operation: "ui.load", origin: self.traceOrigin)
+            let transactionId = SentryPerformanceTracker.shared.startSpan(
+                withName: self.name,
+                nameSource: nameSource,
+                operation: SentrySpanOperation.uiLoad,
+                origin: traceOrigin
+            )
             SentryPerformanceTracker.shared.pushActiveSpan(transactionId)
 
             //According to Apple's documentation, the call to `body` needs to be fast
@@ -83,7 +88,12 @@ public struct SentryTracedView<Content: View>: View {
             }
         }
 
-        let id = SentryPerformanceTracker.shared.startSpan(withName: transactionCreated ? "\(self.name) - body" : self.name, nameSource: nameSource, operation: "ui.load", origin: self.traceOrigin)
+        let id = SentryPerformanceTracker.shared.startSpan(
+            withName: transactionCreated ? "\(self.name) - body" : self.name,
+            nameSource: nameSource,
+            operation: SentrySpanOperation.uiLoad,
+            origin: traceOrigin
+        )
 
         SentryPerformanceTracker.shared.pushActiveSpan(id)
         defer {