+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Internal
 
 - Change macros TEST and TESTCI to SENTRY_TEST and SENTRY_TEST_CI (#4712)
+- Convert constants SentrySpanOperation to Swift ()
 
 ## 8.43.1-beta.0